Evaluation2/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
openai>=1.54.0
